Split SavePersonUseCase to CreatePersonUseCase and UpdatePersonUseCase
