+++ condo-manager-core/condo-manager-core-use-case/src/main/kotlin/com/ikm/condomanager/usecase/person/UpdatePersonUseCase.kt
@@ -1,0 +1,19 @@
+package com.ikm.condomanager.usecase.person
+
+import com.ikm.condomanager.domain.Person
+import com.ikm.condomanager.exception.NotFoundException
+
+/**
+ * Use case for updating a [Person] in the db.
+ */
+interface UpdatePersonUseCase {
+    /**
+     * Updates the given valid [person] in the db.
+     * [Person.id] must be provided.
+     *
+     * @param person the [Person] to be updated.
+     * @return the updated [Person].
+     * @throws NotFoundException if the [Person] to be updated, identified by [Person.id], is not found in the db.
+     */
+    fun update(person: Person): Person
+}
